Extract functionalities for books page into adapter
