+++ lib/kindle_manager.rb
@@ -1,5 +1,6 @@
 require 'amazon_auth'
 require "kindle_manager/version"
+require "kindle_manager/adapters/books_adapter"
 require "kindle_manager/client"
 require "kindle_manager/file_store"
 require "kindle_manager/list_parser"
